pyproject.toml
[tool.poetry]
name = "cern-search-rest-api"
version = "1.4.0-beta"
description = "CERN Search as a Service"
authors = ["Citadel Search <cernsearch.support@cern.ch>"]
license = "MIT"

[tool.poetry.dependencies]
Flask = "^1.1.2"
invenio-access = ">=1.4.1,<1.5.0"
invenio-accounts = ">=1.3.0,<1.4.0"
invenio-admin = ">=1.2.1,<1.3.0"
invenio-app = ">=1.2.6,<1.3.0"
invenio-assets = ">=1.1.3,<1.2.0"
invenio-base = ">=1.2.3,<1.3.0"
invenio-celery = ">=1.2.0,<1.3.0"
invenio-config = ">=1.0.3,<1.1.0"
invenio-db = {version = ">=1.0.5,<1.1.0", extras = ["postgresql", "versioning"]}
invenio-files-processor = {extras = ["tika"], version = "^0.1.0"}
invenio-files-rest = ">=1.2.0,<1.3.0"
invenio-i18n = "<1.3.0,>=1.2.0"
invenio-indexer = ">=1.1.1,<1.2.0"
invenio-jsonschemas = ">=1.1.0,<1.2.0"
invenio-logging = {version = ">=1.3.0,<1.4.0", extras = ["sentry-sdk"]}
invenio-oauth2server = ">=1.2.0,<1.3.0"
invenio-oauthclient = "<1.4.0,>=1.3.5"
invenio-records = {version = ">=1.3.1,<1.4.0", extras = ["postgresql"]}
invenio-records-files = ">=1.2.1,<1.3.0"
invenio-records-rest = ">=1.7.1,<1.8.0"
invenio-rest = ">=1.2.1,<1.3.0"
invenio-search = {version = ">=1.3.1,<1.4.0", extras = ["elasticsearch7"]}
invenio-theme = ">=1.1.4,<1.2.0"
python = "^3.6.1"
uWSGI = "^2.0.19"
uwsgi-tools = "^1.1.1"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.4"
flake8-docstrings = "^1.5.0"
isort = "^5.6.4"
pytest-dotenv = "^0.5.2"
pytest-invenio = "^1.4.0"
pytest-mock = "^3.3.1"
flake8-logging-format = "^0.6.0"
pre-commit = "^2.9.3"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools"]
build-backend = "poetry.core.masonry.api"
